+++ package.json
@@ -5,7 +5,8 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "lint": "eslint ./src/"
+    "lint": "eslint ./src/",
+    "test": "jest"
   },
   "dependencies": {
     "on-covid-19": "^0.1.3"
