+++ src/estimator.js
@@ -1,75 +1,9 @@
-// input
-// {
-//     region: {
-//         name: "Africa",
-//         avgAge: 19.7,
-//         avgDailyIncomeInUSD: 5,
-//         avgDailyIncomePopulation: 0.71
-//     },
-//     periodType: "days",
-//     timeToElapse: 58,
-//     reportedCases: 674,
-//     population: 66622705,
-//     totalHospitalBeds: 1380614
-// }
-
-// output
-// {
-//     data: {},
-//     impact: {},
-//     severeImpact: {}
-//     // the input data you got
-//     // your best case estimation
-//     // your severe case estimation
-// }
-
-// const getReportedCases = (reportedCase)
-
-const getRequestedTimeInDays = (periodType, timeToElapse) => {
-  let numberOfDays;
-
-  switch (periodType) {
-    case 'days':
-      numberOfDays = timeToElapse;
-      break;
-
-    case 'weeks':
-      numberOfDays = timeToElapse * 7;
-      break;
-
-    case 'months':
-      numberOfDays = timeToElapse * 30;
-      break;
-
-    default:
-      throw new Error(`Invalid Period Type ${periodType}`);
-  }
-
-  return Math.trunc(numberOfDays);
-};
-
-const getInfectionsByRequestedTime = (
-  currentlyInfected,
-  requestedTimeInDays
-) => {
-  const factor = Math.trunc(requestedTimeInDays / 3);
-  return currentlyInfected * 2 ** factor;
-};
-
-const getHospitalBedsByRequestedTime = (
-  totalHospitalBeds,
-  severeCasesByRequestedTime
-) => {
-  const availableBeds = Math.trunc(0.35 * totalHospitalBeds);
-  return availableBeds - severeCasesByRequestedTime;
-};
-
-const getDollarsInFlight = (
-  infectionsByRequestedTime,
-  population,
-  income,
-  days
-) => infectionsByRequestedTime * population * income * days;
+import {
+  getRequestedTimeInDays,
+  getInfectionsByRequestedTime,
+  getHospitalBedsByRequestedTime,
+  getDollarsInFlight
+} from './utils';
 
 const calculateCovid19Impact = (data) => {
   const currentlyInfected = data.reportedCases * 10;
